Adjust the package management for the daemon
The logentries-daemon package tries to start the agent when you install it,
this doesn't work with Chef as we haven't had opportunity to set the host
configuration yet.

Change the way the packages are managed so the commandline tools are installed
immediately, and then the daemon install is delayed until the end of the Chef
run. This way we get the chance to execute the host configuration before installing
the daemon.